Charts: Extract resolveFontSize into a shared utility
resolveFontSize was defined locally inside use-chart-margin and then
re-exported through the hooks barrel so pie-chart could reuse it for
label sizing. Now that there are multiple call sites, move it into its
own file under src/utils alongside the other chart helpers and export
it from the utils barrel.

Drop the re-export from the hooks barrel and update pie-chart to import
directly from utils.
