+++ projects/js-packages/charts/src/hooks/index.ts
@@ -2,7 +2,7 @@ export { useDeepMemo } from './use-deep-memo';
 export { useChartMouseHandler } from './use-chart-mouse-handler';
 export { useXYChartTheme } from './use-xychart-theme';
 export { useChartDataTransform } from './use-chart-data-transform';
-export { useChartMargin, resolveFontSize } from './use-chart-margin';
+export { useChartMargin } from './use-chart-margin';
 export { useElementSize } from './use-element-size';
 export { useTextTruncation } from './use-text-truncation';
 export { useZeroValueDisplay } from './use-zero-value-display';
