+++ projects/js-packages/charts/src/utils/index.ts
@@ -26,3 +26,6 @@ export * from './color-utils';
 
 // CSS utilities
 export { resolveCssVariable } from './resolve-css-var';
+
+// Font sizing utilities
+export { resolveFontSize } from './resolve-font-size';
